+++ NEWS.md
@@ -7,6 +7,7 @@ there is even/first/default parts).
 - fix behavior regression with headers and footers replace_text_at_bkm. 
 The previous behavior "don't trigger an error if a bookmark is missing" 
 is back. 
+- force Word default continuous section if no type specified
 
 ## Changes